Fix some tests. More need fixing.
